Don't try to remove /opt/TurboVNC/lib unless it exists, i.e. unless uninstalling TurboVNC 0.6 or prior
git-svn-id: svn+ssh://svn.code.sf.net/p/turbovnc/code/trunk@1249 799e4f7b-5fd2-41f6-823c-2ecc41bc7f0b

+++ vnc_unixsrc/uninstall.sh
@@ -45,7 +45,7 @@ cd $PWD
 
 echo Removing directories ...
 rmdir /opt/TurboVNC/bin 2>&1 || RETCODE=-1
-rmdir /opt/TurboVNC/lib 2>&1 || RETCODE=-1
+if [ -d /opt/TurboVNC/lib ]; then rmdir /opt/TurboVNC/lib 2>&1 || RETCODE=-1; fi
 rmdir /opt/TurboVNC/man/man1 2>&1 || RETCODE=-1
 rmdir /opt/TurboVNC/man 2>&1 || RETCODE=-1
 rmdir /opt/TurboVNC 2>&1 || RETCODE=-1
